LayoutTests/imported/mozilla/css-animations/test_animations-dynamic-changes-expected.txt


PASS Animations preserve their startTime when changed 
PASS Updated Animations maintain their order in the list 
FAIL Only the startTimes of existing animations are preserved assert_true: New Animation is prepended to start of list expected true got false
FAIL Animations are removed from the start of the list while preserving the state of existing Animations assert_equals: List of Animations was trimmed expected 1 but got 26
FAIL Animation state is preserved when interleaving animations in list assert_not_equals: New Animations are added to start of the list got disallowed value object "[object CSSAnimation]"
